rest_app/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0